ref: initを別ファイルに記載
